# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Minha API",
    "description": "Descrição da minha API"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Rota principal",
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "summary": "Obtém detalhes do usuário por ID",
        "parameters": [{
          "in": "path",
          "name": "id",
          "required": true,
          "description": "ID do usuário",
          "schema": {
            "type": "string"
          }
        }],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operação bem-sucedida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "401": {
            "description": "Acesso negado!"
          }
        }
      }
    },
    
    "/auth/register": {
      "post": {
        "summary": "Registra um novo usuário",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "confirmpassword": {
                    "type": "string"
                  }
                },
                "required": ["name", "email", "password", "confirmpassword"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso"
          },
          "422": {
            "description": "Erro de validação"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Autenticação bem-sucedida"
          },
          "422": {
            "description": "Erro de validação"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      }
    }
  },
  
    "components": {
      "schemas": {
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "description": "ID do usuário"
            },
            "name": {
              "type": "string",
              "description": "Nome do usuário"
            },
            "email": {
              "type": "string",
              "description": "E-mail do usuário"
            }
          }
        }
      },
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "BearerAuth": []
      }
    ]
  }
  
  